moved most things back to a top level namespace
It occured to me that having users in a top level namespace is
the dramatically more error prone of the two options, i.e., having
users in the top level namespace is kind of a rare thing as far as
websites go, and people are probably more likely to assume that
quirell.net/login gives you the login page instead of the user
called '@login', and thusly more likely to lead to lots of
confusion.

Now what we -can- do is have URLs redirect to
quirell.net/user/<PATH> if quirell.net/<PATH> doesn't exist.

Also... it should be noted that this behavior will probably change
in the future >_<