[mostly] completely unneeded database changes
also added two (yes, just two -_-) attributes to posts

That's what started it actually. I had this idea that since I'll
be adding more things to posts in the future that I should write
code to update all the nodes in the database. And then 6 hours
of work later I decided that all I would need at this point is
a for loop and a key value swap. Sometime in the middle of that
6 hours there was a massive series of "get or set value" lines...

...anways, yea. added two attributes to posts.