user loader now sends a request to the neo4j database
there was an issue on heroku where the list of active users was
disappearing from RAM every for requests. I googled the issue
and found out that I shouldn't have been making user_loader
ask for something from RAM, but instead it is supposed to query
to DB.

I assume I implemented it the original way to cut down on DB
queries, but now that I'm having this (what I assume is a)
disappearing RAM issue (to be explicit: it might not be that!),
it's definitely got to go.